@@ -9,9 +9,10 @@ import 'package:package_info/package_info.dart';
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
 
-  const MethodChannel channel =
+  const channel =
       MethodChannel('plugins.flutter.io/package_info');
-  List<MethodCall> log;
+
+  late List<MethodCall> log;
 
   channel.setMockMethodCallHandler((MethodCall methodCall) async {
     log.add(methodCall);
@@ -34,7 +35,7 @@ void main() {
   });
 
   test('fromPlatform', () async {
-    final PackageInfo info = await PackageInfo.fromPlatform();
+    final info = await PackageInfo.fromPlatform();
     expect(info.appName, 'package_info_example');
     expect(info.buildNumber, '1');
     expect(info.packageName, 'io.flutter.plugins.packageinfoexample');
