@@ -6,13 +6,13 @@ import 'dart:io';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:package_info/package_info.dart';
-import 'package:package_info_example/main.dart';
+import 'package:package_info_example/main.dart' show MyApp;
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
   testWidgets('fromPlatform', (WidgetTester tester) async {
-    final PackageInfo info = await PackageInfo.fromPlatform();
+    final info = await PackageInfo.fromPlatform();
     // These tests are based on the example app. The tests should be updated if any related info changes.
     if (Platform.isAndroid) {
       expect(info.appName, 'package_info_example');
